+++ front/src/components/common/server-component/centerCard.tsx
@@ -3,7 +3,7 @@ import React, { Suspense } from "react";
 import { getCachedPaginatedPosts } from "@/lib/server-only/postLib";
 import { PostType } from "@/atoms/post";
 import PostList from "@/components/posts/view/postList";
-import LoadMorePostCards from "@/components/posts/view/LoadMorePostCards";
+import LoadMorePostCards from "@/components/posts/card/LoadMorePostCards";
 
 
 /*TODO
