+++ front/src/components/posts/card/components/QuestButtons.tsx
