+++ front/src/components/posts/card/components/postThumbnail.tsx
